% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempConversion.R
\name{CtoF}
\alias{CtoF}
\title{Celsius conversion}
\usage{
CtoF(Ctemp)
}
\arguments{
\item{Ctemp}{The temperature in Celsius}
}
\value{
The temperature in degrees Fahrenheit
}
\description{
Convert degrees Celsius temperatures to degrees Fahrenheit
}
\examples{
temp1 <- FtoC(22);
temp2 <- FtoC(c(-2,12,23))
}
