% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempConversion.R
\name{FtoC}
\alias{FtoC}
\title{Fahrenheit conversion}
\usage{
FtoC(Ftemp)
}
\arguments{
\item{Ftemp}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in degrees Celsuis
}
\description{
Convert degrees Fahrenheit temperatures to degrees Celsius
}
\examples{
temp1 <- FtoC(50);
temp2 <- FtoC(c(50,63,23))
}
